requirements.txt
Pypubsub~=4.0.3
StrEnum~=0.4.8
matplotlib~=3.6.0
sv-ttk~=2.0
